+++ Server/StreetEmpire.Api/Data/GameDbContext.cs
@@ -53,6 +53,7 @@ public sealed class GameDbContext(DbContextOptions<GameDbContext> options) : DbC
             entity.Property(x => x.ProfilePronouns).HasMaxLength(64);
             entity.Property(x => x.ProfileLocation).HasMaxLength(64);
             entity.Property(x => x.ProfileAccent).HasConversion<string>().HasMaxLength(16).HasDefaultValue(ProfileAccent.Gold);
+            entity.Property(x => x.ProfileBanner).HasConversion<string>().HasMaxLength(16).HasDefaultValue(ProfileBanner.None);
             entity.Property(x => x.DirectMessagePolicy).HasConversion<string>().HasMaxLength(16);
             entity.Property(x => x.EmailSecurityNotices).HasDefaultValue(true);
             entity.Property(x => x.EmailCombatNotices).HasDefaultValue(true);